updated routes for new nux flow
