+++ aux_ui/lib/named_routing/router.dart
@@ -1,22 +1,26 @@
 import 'package:aux_ui/named_routing/routing_constants.dart';
-import 'package:aux_ui/screens/host_signup/host_queue_confirmation.dart';
-import 'package:aux_ui/screens/host_signup/host_invite.dart';
-import 'package:aux_ui/screens/host_signup/host_name_queue.dart';
-import 'package:aux_ui/screens/host_signup/host_spotify_link.dart';
-import 'package:aux_ui/screens/guest_signup/guest_signup.dart';
-import 'package:aux_ui/screens/guest_signup/guest_join_queue.dart';
-import 'package:aux_ui/screens/alt_nux_intro.dart';
+import 'package:aux_ui/screens/create_queue/host_queue_confirmation.dart';
+import 'package:aux_ui/screens/create_queue/host_invite.dart';
+// import 'package:aux_ui/screens/create_queue/host_name_queue.dart';
+import 'package:aux_ui/screens/link_spotify.dart';
+// import 'package:aux_ui/screens/join_queue/guest_signup.dart';
+import 'package:aux_ui/screens/join_queue/join_queue.dart';
+import 'package:aux_ui/screens/nux_intro.dart';
 import 'package:flutter/material.dart';
 
 Route<dynamic> generateRoute(RouteSettings settings) {
+  Object args = settings.arguments;
   // TODO: confirm where we replace the route and where we push/pop, decide where to have default null next/back
   switch (settings.name) {
     case AuxHomeRoute:
       return MaterialPageRoute(builder: (context) => NuxIntro());
-    case HostSpotifyLinkRoute:
-      return MaterialPageRoute(builder: (context) => HostSpotifyLink(nextPage: HostNameQueueRoute));
-    case HostNameQueueRoute:
-      return MaterialPageRoute(builder: (context) => HostNameQueue(nextPage: HostInviteRoute));
+    
+    case LinkSpotifyRoute:
+      if (args.toString() == 'host') {
+        return MaterialPageRoute(builder: (context) => LinkSpotify(nextPage: HostInviteRoute));
+      }
+      return MaterialPageRoute(builder: (context) => LinkSpotify(nextPage: GuestJoinQueueRoute));
+    
     case HostInviteRoute:
       String queueName = settings.arguments;
       return MaterialPageRoute(builder: (context) => HostInvite(
@@ -24,14 +28,16 @@ Route<dynamic> generateRoute(RouteSettings settings) {
         nextPage: HostConfirmationRoute
         )
       );
+    
     case HostConfirmationRoute:
       return MaterialPageRoute(builder: (context) => HostQueueConfirmation(backPage: HostNameQueueRoute));
-    case GuestSignupRoute:
-      return MaterialPageRoute(builder: (context) => GuestSignup(nextPage: GuestJoinQueueRoute));
+    
     case GuestJoinQueueRoute:
       return MaterialPageRoute(builder: (context) => GuestJoinQueue(nextPage: '/'));
+    
     case NuxIntroRoute:
       return MaterialPageRoute(builder: (context) => NuxIntro());
+    
     default:
       return MaterialPageRoute(builder: (context) => NuxIntro());
   }
