+++ aux_ui/lib/named_routing/routing_constants.dart
@@ -1,8 +1,9 @@
 const String AuxHomeRoute = '/';
-const String NuxIntroRoute = 'alt_nux_intro';
-const String GuestSignupRoute = 'guest_signup';
+const String NuxIntroRoute = 'nux_intro';
+const String LinkSpotifyRoute = 'link_spotify';
+// const String GuestSignupRoute = 'guest_signup';
 const String GuestJoinQueueRoute = 'guest_join_queue';
-const String HostSpotifyLinkRoute = 'host_spotify_link';
+// const String HostSpotifyLinkRoute = 'host_spotify_link';
 const String HostNameQueueRoute = 'host_name_queue';
 const String HostInviteRoute = 'host_invite';
 const String HostConfirmationRoute = 'host_confirmation';
