+++ aux_ui/lib/screens/join_queue/join_queue.dart
