+++ aux_ui/lib/screens/nux_intro.dart
@@ -1,42 +1,50 @@
+
 import 'package:flutter/material.dart';
 import 'package:aux_ui/theme/aux_theme.dart';
-import '../widgets/aux_card.dart';
-import '../widgets/text_half_card.dart';
+import 'package:flutter/rendering.dart';
+import 'package:aux_ui/widgets/nux_container.dart';
+import 'package:aux_ui/widgets/buttons/icon_bar_button.dart';
+import 'package:aux_ui/named_routing/routing_constants.dart';
 
 class NuxIntro extends StatelessWidget {
+
+  void _signup(BuildContext ctx, String userType) {
+    Navigator.pushNamed(ctx, LinkSpotifyRoute, arguments: userType);
+  }
+
   @override
   Widget build(BuildContext context) {
-    SizeConfig().init(context);
-    return Container(
-        height: SizeConfig.safeBlockVertical * 100,
-        width: SizeConfig.safeBlockHorizontal * 100,
-        child: Padding(
-            padding: SizeConfig.notchPadding,
-            child: AuxCard(
-                borderColor: auxAccent,
-                padding: 0.0,
-                child: Stack(children: <Widget>[
-                  Column(
-                    mainAxisSize: MainAxisSize.max,
-                    children: <Widget>[
-                      Expanded(
-                          child: TextHalfCard(
-                              orientation: "top",
-                              background: auxPrimary,
-                              textStyle: auxDisp3,
-                              text: ">> host\nan aux queue")),
-                      Expanded(
-                          child: TextHalfCard(
-                              orientation: "bottom",
-                              background: auxAccent,
-                              textStyle: auxDisp3Inv,
-                              text: ">> join\nan aux queue"))
-                    ],
-                  ),
-                  Center(child: Padding(
-                      padding: EdgeInsets.all(20),
-                      child: Text('or', style: auxDispMidAccent)
-                  )),
-                ]))));
+    return 
+      NuxContainer(
+        title: 'aux', 
+        topWidget: 
+          Align(
+            alignment: Alignment.bottomLeft,
+            child: Text("let's get started", style: auxDisp3),
+          ),
+        bottomWidget:  
+          Align(
+            alignment: Alignment.topCenter,
+              child: Column(
+                children: <Widget>[
+                  ButtonTheme(
+                      minWidth: double.infinity,
+                      child: IconBarButton(
+                          text: 'host a new queue',
+                          onPressed: () => _signup(context, 'host'),
+                        ),
+                    ),
+                    Padding(padding: EdgeInsets.all(18)),
+                    ButtonTheme(
+                      minWidth: double.infinity,
+                      child: IconBarButton(
+                          text: 'join an existing queue',
+                          onPressed: () => _signup(context, 'guest'),
+                        ),
+                    ),
+                ]
+              )
+          )
+      );
   }
 }
